+++ Core/ALife.Core/Collision/CollisionDetector.cs
@@ -114,11 +114,11 @@ namespace ALife.Core.Collision
             if(PointCircleCollision(sector.CentrePoint, circle)) return true;
 
             //Check the left Geometry.Shapes.Point of the sector
-            Geometry.Shapes.Point leftPoint = GeometryMaths.TranslateByVector(sector.CentrePoint, sector.Orientation, sector.Radius);
+            Geometry.Shapes.Point leftPoint = GeometryMath.TranslateByVector(sector.CentrePoint, sector.Orientation, sector.Radius);
             if(PointCircleCollision(leftPoint, circle)) return true;
 
             //Check the right Geometry.Shapes.Point of the sector
-            Geometry.Shapes.Point rightPoint = GeometryMaths.TranslateByVector(sector.CentrePoint, (sector.Orientation + sector.SweepAngle), sector.Radius);
+            Geometry.Shapes.Point rightPoint = GeometryMath.TranslateByVector(sector.CentrePoint, (sector.Orientation + sector.SweepAngle), sector.Radius);
             if(PointCircleCollision(rightPoint, circle)) return true;
 
             //Now we're checking the line segment collisions
@@ -413,7 +413,7 @@ namespace ALife.Core.Collision
 
         private static bool IsPointWithinSweep(Geometry.Shapes.Point targetPoint, Sector sector)
         {
-            double angleBetweenPoints = GeometryMaths.AngleBetweenPoints(targetPoint, sector.CentrePoint);
+            double angleBetweenPoints = GeometryMath.AngleBetweenPoints(targetPoint, sector.CentrePoint);
             Angle abp = new Angle(angleBetweenPoints, true);
 
             //Angle minimum = new Angle(0);
