+++ Core/ALife.Core/Geometry/Shapes/ChildShapes/ChildSector.cs
@@ -56,7 +56,7 @@ namespace ALife.Core.Geometry.Shapes.ChildShapes
             get
             {
                 Angle startAngle = Parent.Orientation + OrientationAroundParent;
-                return GeometryMaths.TranslateByVector(Parent.CentrePoint, startAngle, distanceFromParentCentre);
+                return GeometryMath.TranslateByVector(Parent.CentrePoint, startAngle, distanceFromParentCentre);
             }
             set
             {
