+++ Core/ALife.Core/WorldObjects/Agents/AgentActions/MoveCluster.cs
@@ -65,7 +65,7 @@ namespace ALife.Core.WorldObjects.Agents.AgentActions
             string stop = "Stop" + direction;
 
             double dirValue = SubActions[go].Intensity - SubActions[stop].Intensity;
-            dirValue = ExtraMaths.Clamp(dirValue, 0.0, 1.0);
+            dirValue = ExtraMath.Clamp(dirValue, 0.0, 1.0);
             return dirValue;
         }
 
@@ -81,13 +81,13 @@ namespace ALife.Core.WorldObjects.Agents.AgentActions
             forwardDist = Speed * forwardMagnitude;
             rightDist = Speed * rightMagnitude;
 
-            Geometry.Shapes.Point tempPoint = GeometryMaths.TranslateByVector(origin, theShape.Orientation, forwardDist);
-            Geometry.Shapes.Point finalPoint = GeometryMaths.TranslateByVector(tempPoint, theShape.Orientation.Radians + (Math.PI / 2), rightDist);
+            Geometry.Shapes.Point tempPoint = GeometryMath.TranslateByVector(origin, theShape.Orientation, forwardDist);
+            Geometry.Shapes.Point finalPoint = GeometryMath.TranslateByVector(tempPoint, theShape.Orientation.Radians + (Math.PI / 2), rightDist);
 
             double halfXLength = theShape.BoundingBox.XLength / 2;
             double halfYHeight = theShape.BoundingBox.YHeight / 2;
-            finalPoint.X = ExtraMaths.Clamp(finalPoint.X, halfXLength, Planet.World.WorldWidth - halfXLength);
-            finalPoint.Y = ExtraMaths.Clamp(finalPoint.Y, halfYHeight, Planet.World.WorldHeight - halfYHeight);
+            finalPoint.X = ExtraMath.Clamp(finalPoint.X, halfXLength, Planet.World.WorldWidth - halfXLength);
+            finalPoint.Y = ExtraMath.Clamp(finalPoint.Y, halfYHeight, Planet.World.WorldHeight - halfYHeight);
 
             theShape.CentrePoint = finalPoint;
