+++ Core/ALife.Core/WorldObjects/Agents/Properties/PropertyInput.cs
@@ -68,14 +68,14 @@ namespace ALife.Core.WorldObjects.Agents.Properties
 
         public void ChangePropertyTo(double value)
         {
-            Value = ExtraMaths.Clamp(value, PropertyMinimum, PropertyMaximum);
+            Value = ExtraMath.Clamp(value, PropertyMinimum, PropertyMaximum);
             modified = true;
         }
 
         public override double Value
         {
             get { return base.Value; }
-            set { base.Value = ExtraMaths.Clamp(value, PropertyMinimum, PropertyMinimum); }
+            set { base.Value = ExtraMath.Clamp(value, PropertyMinimum, PropertyMinimum); }
 
         }
