+++ Core/ALife.Core/WorldObjects/Agents/Senses/GoalSenseCluster.cs
@@ -125,9 +125,9 @@ namespace ALife.Core.WorldObjects.Agents.Senses
                 double yTarg = myCP.Y < targBB.MinY ? targBB.MinY : targBB.MaxY;
                 Point target = new Point(xTarg, yTarg);
 
-                distanceValue = (int)GeometryMaths.DistanceBetweenTwoPoints(target, myCP);
+                distanceValue = (int)GeometryMath.DistanceBetweenTwoPoints(target, myCP);
 
-                double angleBetweenPoints = GeometryMaths.AngleBetweenPoints(target, myCP);
+                double angleBetweenPoints = GeometryMath.AngleBetweenPoints(target, myCP);
                 Angle abp = new Angle(angleBetweenPoints, true);
                 rotationValue = CalculateRotationFrom((int)abp.Degrees);
             }
@@ -150,10 +150,10 @@ namespace ALife.Core.WorldObjects.Agents.Senses
             Point myCP = myShape.CentrePoint;
             Point target = new Point(targetCircle.CentrePoint.X, targetCircle.CentrePoint.Y);
 
-            int distanceValue = (int)GeometryMaths.DistanceBetweenTwoPoints(target, myCP);
+            int distanceValue = (int)GeometryMath.DistanceBetweenTwoPoints(target, myCP);
             distanceValue -= (int)targetCircle.Radius;
 
-            double angleBetweenPoints = GeometryMaths.AngleBetweenPoints(target, myCP);
+            double angleBetweenPoints = GeometryMath.AngleBetweenPoints(target, myCP);
             Angle abp = new Angle(angleBetweenPoints, true);
             int rotationValue = CalculateRotationFrom((int)abp.Degrees);
