+++ Core/ALife.Core/WorldObjects/Prebuilt/Wall.cs
@@ -76,7 +76,7 @@ namespace ALife.Core.WorldObjects.Prebuilt
                 Angle ori = wall.Shape.Orientation.Clone();
                 double indexer = i - ((numSplits + 1) / 2.0);
 
-                Geometry.Shapes.Point p = GeometryMaths.TranslateByVector(wall.Shape.CentrePoint, ori, segmentLength * indexer);
+                Geometry.Shapes.Point p = GeometryMath.TranslateByVector(wall.Shape.CentrePoint, ori, segmentLength * indexer);
                 Wall w = new Wall(p, segmentLength, ori, wall.IndividualLabel + "~" + (i + 1));
                 segments.Add(w);
             }
